+++ src/chromosome/bit.rs
@@ -8,7 +8,6 @@ pub struct Bit {
     pub fitness_score: Option<FitnessValue>,
     pub genes_hash: Option<GenesHash>,
     pub age: usize,
-    pub reference_id: usize,
 }
 
 impl super::Chromosome for Bit {
@@ -40,13 +39,11 @@ impl super::Chromosome for Bit {
         self.age = 0;
         self.fitness_score = None;
         self.genes_hash = genes_hash;
-        self.reference_id = usize::MAX;
     }
     fn copy_state(&mut self, other: &Self) {
         self.age = other.age;
         self.fitness_score = other.fitness_score;
         self.genes_hash = other.genes_hash;
-        self.reference_id = other.reference_id;
     }
 }
 impl super::GenesOwner for Bit {
@@ -57,7 +54,6 @@ impl super::GenesOwner for Bit {
             fitness_score: None,
             genes_hash: None,
             age: 0,
-            reference_id: usize::MAX,
         }
     }
     fn genes(&self) -> &FixedBitSet {